+++ tests/integrations/test_integration_base_yaml.py
@@ -382,6 +382,7 @@ class YamlIntegrationTests:
         "implement",
         "plan",
         "checklist",
+        "revise",
         "specify",
         "tasks",
         "taskstoissues",